make backend work with minion 7
This does not implement all the features of Minion 7, but is a bare
minimum to make the backend work. This implements:

* worker commands
* auto-retry attempts

This also has some fixes to make the backend work on MySQL 5.6.4 and
later, which made `timestamp` columns much more strict about what they
accept.

What remains to be implemented:

* parent jobs (dependencies)
* locks

and maybe some other things, but those are the big two...