Require at least MySQL version 5.6.5
The reason for this requirement is that the current codebase requires
more than one TIMESTAMP column in a table which uses either one of the
DEFAULT CURRENT_TIMESTAMP and ON UPDATE CURRENT_TIMESTAMP clauses.  This
was a restriction in MySQL which was [removed in version
5.6.5](https://dev.mysql.com/doc/relnotes/mysql/5.6/en/news-5-6-5.html#mysqld-5-6-5-data-types).
It seems that the versions of MariaDB after the 5.5 series (i.e.  10.x
and above) don't have this restriction, hence requiring a server version
number greater than or equal to 5.6.5 is sufficient to ensure that the
code works with the available MySQL server installation.  A somewhat
confusing thing is that the MariaDB 5.5 and 10.0 series seem to be
compatible with the 5.5 series of MySQL as noted on the [MariaDB versus
MySQL Compatibility
page](https://mariadb.com/kb/en/library/mariadb-vs-mysql-compatibility/),
however the 10.0 series doesn't show the TIMESTAMP restriction.
Nevertheless, this change should guard against users trying to use a
version of MySQL which doesn't support the required TIMESTAMP
functionality.