Test MySQL and various MariaDB versions on Travis
In order to do this, one needs to create a build matrix for Travis.
Usually, a build matrix is automatically constructed by Travis, however
only for combinations of Runtime, Environment and Exclusions/Inclusions
(see [Customizing the
Build](https://docs.travis-ci.com/user/customizing-the-build/) for more
details).  Hence it's necessary to construct the build matrix
explicitly.  It turns out that MySQL is currently version 5.6 on Travis
and that they don't support other versions explicitly.  However, they do
support MariaDB in various versions (5.5, 10.0, and 10.1), hence why one
uses the `addons` feature in order to do this.  Each desired
perl-mysql/mariadb combination thus needs to be specified explicitly,
hence this change is rather large and seems quire repetitive.  The
`libmariadbclient-dev` package is required for MariaDB builds so that
the `mysql_config` program is available for `DBD::mysql` during its
build and installation phase.