Replace SELECT FOR UPDATE  then UPDATE with atomic UPDATE
Do an atomic UPDATE that replaces the need for doing a SELECT FOR
UPDATE followed by an UPDATE within a transaction, which should reduce
the possibility of lock wait timeouts.

MySQL doesn't like it if you JOIN to the same table you are updating,
but doesn't mind if you do a sub select on that table and then JOIN to
that instead.