fix deadlocks by removing assignment to id field
The old `UPDATE ...` query contained a useless assignment to the `id` field:
`job.id = @dequeued_job_id := job.id`. The goal was the side-effect of
setting the `@dequeued_job_id` variable that was used later to get the
job information. However, this requires obtaining an exclusive lock on
the index for the `id` field, which cannot be done when there is also
a shared lock (like the one in the dependent subquery for the job's
parents...)

MySQL has a construct for doing an "atomic" `SELECT` followed by
`UPDATE`: `SELECT ... FOR UPDATE`. This locks the returned rows with
exclusive locks so other queries can't read them while we then come back
and perform the update.

Thanks @srchulo for the help debugging this issue, and Grant Street
Group for additional debugging information.

Fixes #28