fix note values not accepting UTF-8 in the JSON
JSON should be in UTF-8, but also we shouldn't care whatsoever what
encoding it is: MySQL's `JSON` type is a binary type not a character
type. We're supporting MySQLs from before the `JSON` type, but we should
still be using a binary type here.

Thanks @uralm1 for reporting this issue and suggesting the fix!

Fixes #32