remove transaction from dequeue query
We have a SELECT and then an UPDATE, which has a race condition. So, we
check to see if the UPDATE modified anything. If it did not, we try to
get another job.

Thanks @larryl for the original SQL for this. I added the HAVING clause
to fix the case of a job with more than one parent job.