fix removed parent blocking child job
When a parent job is removed, or any other way a job could depends on
a job that doesn't exist, the job should not be held up. There was
a bug, though, where if a parent job was removed, the corresponding
records in the depends table were not cleaned up. Since adding
the state information to the depends table, it was possible for these
extraneous rows to cause child jobs to be stuck until the database was
fixed manually.

Thanks @larryl of @GrantStreetGroup for reporting this issue!