prefetch jobs to improve performance for short jobs
This fixed CPAN Testers processing to finally have more than a single
"active" job. Literally the lookup of the next job took longer than the
job itself, so the worker spent most of its time waiting for the dequeue
query to come back.

Now since we go through all the effort of joining, sorting, and
filtering, we pull more than a single row from the resultset and
pull from that as needed.